disable mozilla image smoothing
since the unit tests are now client-side, we can do this. note how i take
extra care to disable it for the drawing surface.

i also added a small optimization: getContext('2d') is now cached
in the surface sicne i'm not sure if FF will remember the setting
of mozImageSmoothing.

also fixed unit tests for this change