reactivate image-smoothing; nicer code for smooth(), noSmooth()
there are now hidden functions _noSmooth() and _smooth() on the
Surface which can be used to disable / enable anti-aliasing and filtering
on most browsers that support it.

i re-activated it per default because it really is a lot prettier. fixed the unit
tests again to work for those changes :|

also undid an accidentle change in gamejs.display (style.width) which
would break mouse coordinates in events.