greatly improve QUnit.surfaceEqual
instead of comparing the dataimg strings, we know greate a fresh
test canvas and blit the two images on top with the composition activated
as 'xor'. this way, only the content which is present in only one of the two
pictures is drawn. if this resulting content is less than 8% (tuned number)
we accept the two images as equal