remove gamejs-wrapped
i will in a following commit ignore it. it will only be present in gamejs
releases but not in the git development version. if one wants to work
with the git version, this file has to be rebuild.

it's just to messy to keep it updated all the time and hopefully convinient
enough to have it pre-built in a release version.